requirements.txt
numpy
scikit-learn
scipy
Pillow
flask
matplotlib
